Moved exceptions
Exceptions are now all put in a single file. Easier to look through
later.
